Use a different Lua func to toggle display layers
This change to the Lua script accommodates a corresponding change in the
custom modifications made to DeSmuME. It also updates README.md to reflect
those changes, and to contain the full diff of the DeSmuME source.

The Lua function I originally added was gui.togglelayer, which simply
called TwiddleLayer from display.h to toggle the state of a single layer.
I've replaced this DeSmuME modification with one that takes a slightly
different approach: instead it defines a function called gui.setlayermask,
which takes two bitfields representing the desired layer states for each
screen. Since display.h and its associated layer-visiblity-toggling logic
is specific to the Windows frontend, this new change also ensures that the
modification only takes effect on the Windows build of DeSmuME, meaning
that gui.setlayermask is unsupported on MacOS and Linux.